Create UniqueByFunction super class
